+++ earthaccess/store.py
@@ -15,17 +15,41 @@ import requests
 import s3fs
 from multimethod import multimethod as singledispatchmethod
 from pqdm.threads import pqdm
+from tenacity import (
+    retry,
+    retry_if_exception_type,
+    stop_after_attempt,
+    wait_exponential,
+)
 from typing_extensions import deprecated
 
 import earthaccess
 
 from .auth import Auth, SessionWithHeaderRedirection
 from .daac import DAAC_TEST_URLS, find_provider
-from .exceptions import DownloadFailure
 from .results import DataGranule
 from .search import DataCollections
 
 logger = logging.getLogger(__name__)
+
+
+def _is_interactive() -> bool:
+    """Detect if earthaccess is being used in an interactive session.
+    Interactive sessions include Jupyter Notebooks, IPython REPL, and default Python REPL.
+    """
+    try:
+        from IPython import get_ipython  # type: ignore
+
+        # IPython Notebook or REPL:
+        if get_ipython() is not None:
+            return True
+    except ImportError:
+        pass
+
+    import sys
+
+    # Python REPL
+    return hasattr(sys, "ps1")
 
 
 class EarthAccessFile(io.IOBase):
@@ -75,24 +99,57 @@ class EarthAccessFile(io.IOBase):
         return repr(self.f)
 
 
+def _optimal_fsspec_block_size(file_size: int) -> int:
+    """Determine the optimal block size based on file size.
+    Note: we could even be smarter if we know the chunk sizes of the variables
+    we need to cache, e.g. using the `dmrpp` file and the `wellknownparts` cache type.
+
+    Uses `blockcache` for all files with block sizes adjusted by file size:
+
+    - <100MB: 4MB
+    - >100MB: 4–16MB
+
+    Parameters:
+        file_size (int): Size of the file in bytes.
+
+    Returns:
+        block_size (int): Optimal block size in bytes.
+    """
+    if file_size < 100 * 1024 * 1024:
+        block_size = 4 * 1024 * 1024
+    elif 100 * 1024 * 1024 <= file_size < 1024 * 1024 * 1024:
+        block_size = 8 * 1024 * 1024
+    else:
+        block_size = 16 * 1024 * 1024
+
+    return block_size
+
+
 def _open_files(
     url_mapping: Mapping[str, Union[DataGranule, None]],
     fs: fsspec.AbstractFileSystem,
     *,
     pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+    open_kwargs: Optional[Dict[str, Any]] = None,
 ) -> List[fsspec.spec.AbstractBufferedFile]:
     def multi_thread_open(data: tuple[str, Optional[DataGranule]]) -> EarthAccessFile:
         url, granule = data
-        return EarthAccessFile(fs.open(url), granule)  # type: ignore
+        f_size = fs.info(url)["size"]
+        default_cache_type = "background"  # block cache with background fetching
+        default_block_size = _optimal_fsspec_block_size(f_size)
 
-    pqdm_kwargs = {
-        "exception_behaviour": "immediate",
-        "n_jobs": 8,
-        "disable": True,
-        **(pqdm_kwargs or {}),
-    }
+        open_kw = (open_kwargs or fsspec.config.conf or {}).copy()
 
-    return pqdm(url_mapping.items(), multi_thread_open, **pqdm_kwargs)
+        open_kw.setdefault("cache_type", default_cache_type)
+        open_kw.setdefault("block_size", default_block_size)
+
+        f = fs.open(url, **open_kw)
+        return EarthAccessFile(f, granule)  # type: ignore
+
+    # this {#n_jobs} is for the unittests as this method is not public and pqdm will have values at this point
+    return pqdm(
+        url_mapping.items(), multi_thread_open, **(pqdm_kwargs or {"n_jobs": 8})
+    )
 
 
 def make_instance(
@@ -360,7 +417,10 @@ class Store(object):
         granules: Union[List[str], List[DataGranule]],
         provider: Optional[str] = None,
         *,
+        show_progress: Optional[bool] = None,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+        open_kwargs: Optional[Dict[str, Any]] = None,
     ) -> List[fsspec.spec.AbstractBufferedFile]:
         """Returns a list of file-like objects that can be used to access files
         hosted on S3 or HTTPS by third party libraries like xarray.
@@ -369,14 +429,34 @@ class Store(object):
             granules: a list of granule instances **or** list of URLs, e.g. `s3://some-granule`.
                 If a list of URLs is passed, we need to specify the data provider.
             provider: e.g. POCLOUD, NSIDC_CPRD, etc.
+            show_progress: whether or not to display a progress bar. If not specified, defaults to `True` for interactive sessions
+                (i.e., in a notebook or a python REPL session), otherwise `False`.
+            credentials_endpoint: S3 credentials endpoint
             pqdm_kwargs: Additional keyword arguments to pass to pqdm, a parallel processing library.
                 See pqdm documentation for available options. Default is to use immediate exception behavior.
+            open_kwargs: Additional keyword arguments to pass to `fsspec.open`, such as `cache_type` and `block_size`.
+                Defaults to using `blockcache` with a block size determined by the file size (4 to 16MB).
 
         Returns:
-            A list of "file pointers" to remote (i.e. s3 or https) files.
+            A list of "file pointers" to remote (i.e. `s3://` or `https://`) files.
         """
+        if show_progress is None:
+            show_progress = _is_interactive()
+
+        pqdm_kwargs = {
+            "exception_behaviour": "immediate",
+            "n_jobs": 8,
+            "disable": not show_progress,
+            **(pqdm_kwargs or {}),
+        }
         if len(granules):
-            return self._open(granules, provider, pqdm_kwargs=pqdm_kwargs)
+            return self._open(
+                granules,
+                provider,
+                credentials_endpoint=credentials_endpoint,
+                pqdm_kwargs=pqdm_kwargs,
+                open_kwargs=open_kwargs,
+            )
         return []
 
     @singledispatchmethod
@@ -385,7 +465,9 @@ class Store(object):
         granules: Union[List[str], List[DataGranule]],
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+        open_kwargs: Optional[Dict[str, Any]] = None,
     ) -> List[Any]:
         raise NotImplementedError("granules should be a list of DataGranule or URLs")
 
@@ -395,7 +477,9 @@ class Store(object):
         granules: List[DataGranule],
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+        open_kwargs: Optional[Dict[str, Any]] = None,
     ) -> List[Any]:
         fileset: List = []
         total_size = round(sum([granule.size() for granule in granules]) / 1024, 2)
@@ -410,8 +494,9 @@ class Store(object):
             if granules[0].cloud_hosted:
                 access = "direct"
                 provider = granules[0]["meta"]["provider-id"]
-                # if the data has its own S3 credentials endpoint, we will use it
-                endpoint = self._own_s3_credentials(granules[0]["umm"]["RelatedUrls"])
+                endpoint = credentials_endpoint or self._own_s3_credentials(
+                    granules[0]["umm"]["RelatedUrls"]
+                )
                 if endpoint is not None:
                     logger.info(f"using endpoint: {endpoint}")
                     s3_fs = self.get_s3_filesystem(endpoint=endpoint)
@@ -429,6 +514,7 @@ class Store(object):
                         url_mapping,
                         fs=s3_fs,
                         pqdm_kwargs=pqdm_kwargs,
+                        open_kwargs=open_kwargs,
                     )
                 except Exception as e:
                     raise RuntimeError(
@@ -437,10 +523,14 @@ class Store(object):
                         f"Exception: {traceback.format_exc()}"
                     ) from e
             else:
-                fileset = self._open_urls_https(url_mapping, pqdm_kwargs=pqdm_kwargs)
+                fileset = self._open_urls_https(
+                    url_mapping, pqdm_kwargs=pqdm_kwargs, open_kwargs=open_kwargs
+                )
         else:
             url_mapping = _get_url_granule_mapping(granules, access="on_prem")
-            fileset = self._open_urls_https(url_mapping, pqdm_kwargs=pqdm_kwargs)
+            fileset = self._open_urls_https(
+                url_mapping, pqdm_kwargs=pqdm_kwargs, open_kwargs=open_kwargs
+            )
 
         return fileset
 
@@ -450,10 +540,12 @@ class Store(object):
         granules: List[str],
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+        open_kwargs: Optional[Dict[str, Any]] = None,
     ) -> List[Any]:
         fileset: List = []
-
+        s3_fs = None
         if isinstance(granules[0], str) and (
             granules[0].startswith("s3") or granules[0].startswith("http")
         ):
@@ -472,26 +564,29 @@ class Store(object):
         if self.in_region and granules[0].startswith("s3"):
             if provider is not None:
                 s3_fs = self.get_s3_filesystem(provider=provider)
-                if s3_fs is not None:
-                    try:
-                        fileset = _open_files(
-                            url_mapping,
-                            fs=s3_fs,
-                            pqdm_kwargs=pqdm_kwargs,
-                        )
-                    except Exception as e:
-                        raise RuntimeError(
-                            "An exception occurred while trying to access remote files on S3. "
-                            "This may be caused by trying to access the data outside the us-west-2 region."
-                            f"Exception: {traceback.format_exc()}"
-                        ) from e
-                else:
-                    logger.info(f"Provider {provider} has no valid cloud credentials")
+            elif credentials_endpoint is not None:
+                s3_fs = self.get_s3_filesystem(endpoint=credentials_endpoint)
+            if s3_fs:
+                try:
+                    fileset = _open_files(
+                        url_mapping,
+                        fs=s3_fs,
+                        pqdm_kwargs=pqdm_kwargs,
+                        open_kwargs=open_kwargs,
+                    )
+                except Exception as e:
+                    raise RuntimeError(
+                        "An exception occurred while trying to access remote files on S3. "
+                        "This may be caused by trying to access the data outside the us-west-2 region. "
+                        f"Exception: {traceback.format_exc()}"
+                    ) from e
+
                 return fileset
             else:
-                raise ValueError(
-                    "earthaccess cannot derive the DAAC provider from URLs only, a provider is needed e.g. POCLOUD"
+                logger.error(
+                    f"An error occurred while trying to retrieve the cloud credentials for provider: {provider}. endpoint: {credentials_endpoint}"
                 )
+                return fileset
         else:
             if granules[0].startswith("s3"):
                 raise ValueError(
@@ -507,6 +602,8 @@ class Store(object):
         provider: Optional[str] = None,
         threads: int = 8,
         *,
+        credentials_endpoint: Optional[str] = None,
+        show_progress: Optional[bool] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Path]:
         """Retrieves data granules from a remote storage system.
@@ -526,8 +623,11 @@ class Store(object):
                 month, and day of the current date, and `UUID` is the last 6 digits
                 of a UUID4 value.
             provider: a valid cloud provider, each DAAC has a provider code for their cloud distributions
+            credentials_endpoint: If provided, this will be used to get S3 credentials
             threads: Parallel number of threads to use to download the files;
                 adjust as necessary, default = 8.
+            show_progress: whether or not to display a progress bar. If not specified, defaults to `True` for interactive sessions
+                (i.e., in a notebook or a python REPL session), otherwise `False`.
             pqdm_kwargs: Additional keyword arguments to pass to pqdm, a parallel processing library.
                 See pqdm documentation for available options. Default is to use immediate exception behavior
                 and the number of jobs specified by the `threads` parameter.
@@ -543,13 +643,23 @@ class Store(object):
             uuid = uuid4().hex[:6]
             local_path = Path.cwd() / "data" / f"{today}-{uuid}"
 
+        if show_progress is None:
+            show_progress = _is_interactive()
+
         pqdm_kwargs = {
+            "exception_behaviour": "immediate",  # should be overridden by pqdm_kwargs if passed
             "n_jobs": threads,
-            "disable": True,
+            "disable": not show_progress,
             **(pqdm_kwargs or {}),
         }
 
-        return self._get(granules, Path(local_path), provider, pqdm_kwargs=pqdm_kwargs)
+        return self._get(
+            granules,
+            Path(local_path),
+            provider,
+            credentials_endpoint=credentials_endpoint,
+            pqdm_kwargs=pqdm_kwargs,
+        )
 
     @singledispatchmethod
     def _get(
@@ -558,6 +668,7 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Path]:
         """Retrieves data granules from a remote storage system.
@@ -582,6 +693,23 @@ class Store(object):
         """
         raise NotImplementedError(f"Cannot _get {granules}")
 
+    @retry(
+        reraise=True,
+        stop=stop_after_attempt(3),
+        wait=wait_exponential(multiplier=1, min=1, max=10),
+        retry=retry_if_exception_type(Exception),
+    )
+    def download_cloud_file(
+        self, s3_fs: fsspec.AbstractFileSystem, file: str, local_path: Path
+    ) -> Path:
+        file_name = local_path / Path(file).name
+        if file_name.exists():
+            return file_name  # Skip if already exists
+
+        s3_fs.get([file], str(local_path), recursive=False)
+        logger.info(f"Downloading: {file_name}")
+        return file_name
+
     @_get.register
     def _get_urls(
         self,
@@ -589,32 +717,36 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Path]:
         data_links = granules
-        downloaded_files: List = []
-        if provider is None and self.in_region and "cumulus" in data_links[0]:
+        s3_fs = s3fs.S3FileSystem()
+        if (
+            provider is None
+            and credentials_endpoint is None
+            and self.in_region
+            and "cumulus" in data_links[0]
+        ):
             raise ValueError(
                 "earthaccess can't yet guess the provider for cloud collections, "
-                "we need to use one from earthaccess.list_cloud_providers()"
+                "we need to use one from `earthaccess.list_cloud_providers()` or if known the S3 credential endpoint"
             )
         if self.in_region and data_links[0].startswith("s3"):
-            logger.info(f"Accessing cloud dataset using provider: {provider}")
-            s3_fs = self.get_s3_filesystem(provider=provider)
+            if credentials_endpoint is not None:
+                logger.info(
+                    f"Accessing cloud dataset using credentials_endpoint: {credentials_endpoint}"
+                )
+                s3_fs = self.get_s3_filesystem(endpoint=credentials_endpoint)
+            elif provider is not None:
+                logger.info(f"Accessing cloud dataset using provider: {provider}")
+                s3_fs = self.get_s3_filesystem(provider=provider)
 
-            # TODO: make this parallel or concurrent
-            for file in data_links:
-                file_name = local_path / Path(file).name
-                if not file_name.exists():
-                    try:
-                        s3_fs.get(file, str(local_path))
-                        logger.info(f"Downloaded: {file_name}")
-                        downloaded_files.append(file_name)
-                    except Exception as e:
-                        msg = f"Failed to download {file!r} to {file_name!r}: {e}"
-                        logger.error(msg)
-                        raise DownloadFailure(msg)
-            return downloaded_files
+            def _download(file: str) -> Union[Path, None]:
+                return self.download_cloud_file(s3_fs, file, local_path)
+
+            results = pqdm(data_links, _download, **pqdm_kwargs)
+            return [r for r in results if r is not None]
 
         else:
             # if we are not in AWS
@@ -629,10 +761,10 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         *,
+        credentials_endpoint: Optional[str] = None,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Path]:
         data_links: List = []
-        downloaded_files: List = []
         provider = granules[0]["meta"]["provider-id"]
         endpoint = self._own_s3_credentials(granules[0]["umm"]["RelatedUrls"])
         cloud_hosted = granules[0].cloud_hosted
@@ -660,19 +792,12 @@ class Store(object):
 
             local_path.mkdir(parents=True, exist_ok=True)
 
-            # TODO: make this async
-            for file in data_links:
-                file_name = local_path / Path(file).name
-                if not file_name.exists():
-                    try:
-                        s3_fs.get(file, str(local_path))
-                        logger.info(f"Downloaded: {file_name}")
-                        downloaded_files.append(file_name)
-                    except Exception as e:
-                        msg = f"Failed to download {file!r} to {file_name!r}: {e}"
-                        logger.error(msg)
-                        raise DownloadFailure(msg)
-            return downloaded_files
+            def _download(file: str) -> Union[Path, None]:
+                return self.download_cloud_file(s3_fs, file, local_path)
+
+            results = pqdm(data_links, _download, **pqdm_kwargs)
+            return [r for r in results if r is not None]
+
         else:
             # if the data are cloud-based, but we are not in AWS,
             # it will be downloaded as if it was on prem
@@ -701,8 +826,14 @@ class Store(object):
             local_thread_session.auth = original_session.auth
             self.thread_locals.local_thread_session = local_thread_session
 
+    @retry(
+        reraise=True,
+        stop=stop_after_attempt(3),
+        wait=wait_exponential(multiplier=1, min=1, max=10),
+        retry=retry_if_exception_type(Exception),
+    )
     def _download_file(self, url: str, directory: Path) -> Path:
-        """Download a single file from an on-prem location, a DAAC data center.
+        """Download a single file using a bearer token.
 
         Parameters:
             url: the granule url
@@ -717,23 +848,18 @@ class Store(object):
         local_filename = url.split("/")[-1]
         path = directory / Path(local_filename)
         if not path.exists():
-            try:
-                original_session = self.get_requests_session()
-                # This reuses the auth cookie, we make sure we only authenticate N threads instead
-                # of one per file, see #913
-                self._clone_session_in_local_thread(original_session)
-                session = self.thread_locals.local_thread_session
-                with session.get(url, stream=True, allow_redirects=True) as r:
-                    r.raise_for_status()
-                    with open(path, "wb") as f:
-                        # Cap memory usage for large files at 1MB per write to disk per thread
-                        # https://docs.python-requests.org/en/latest/user/quickstart/#raw-response-content
-                        for chunk in r.iter_content(chunk_size=1024 * 1024):
-                            f.write(chunk)
-            except Exception as e:
-                msg = f"Failed to download {url!r} to {path!r}: {e}"
-                logger.error(msg)
-                raise DownloadFailure(msg)
+            original_session = self.get_requests_session()
+            # This reuses the auth cookie, we make sure we only authenticate N threads instead
+            # of one per file, see #913
+            self._clone_session_in_local_thread(original_session)
+            session = self.thread_locals.local_thread_session
+            with session.get(url, stream=True, allow_redirects=True) as r:
+                r.raise_for_status()
+                with open(path, "wb") as f:
+                    # Cap memory usage for large files at 1MB per write to disk per thread
+                    # https://docs.python-requests.org/en/latest/user/quickstart/#raw-response-content
+                    for chunk in r.iter_content(chunk_size=1024 * 1024):
+                        f.write(chunk)
         else:
             logger.info(f"File {local_filename} already downloaded")
         return path
@@ -783,11 +909,14 @@ class Store(object):
         url_mapping: Mapping[str, Union[DataGranule, None]],
         *,
         pqdm_kwargs: Optional[Mapping[str, Any]] = None,
+        open_kwargs: Optional[Dict[str, Any]] = None,
     ) -> List[fsspec.AbstractFileSystem]:
         https_fs = self.get_fsspec_session()
 
         try:
-            return _open_files(url_mapping, https_fs, pqdm_kwargs=pqdm_kwargs)
+            return _open_files(
+                url_mapping, https_fs, pqdm_kwargs=pqdm_kwargs, open_kwargs=open_kwargs
+            )
         except Exception:
             logger.exception(
                 "An exception occurred while trying to access remote files via HTTPS"