+++ CHANGELOG.md
@@ -9,8 +9,14 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ### Changed
 
-- By default, _disable_ progress bars during downloading.
+- Change default cache behavior in fsspec from `readahead` to `blockcache`.
+  Allow user defined config with `open_kwargs` in the `.open()` method.
+  This improves performance by an order of magnitude.
+  ([#251](https://github.com/nsidc/earthaccess/discussions/251))([#771](https://github.com/nsidc/earthaccess/discussions/771))
+  ([@betolink](https://github.com/betolink))
+- Add `show_progress` argument to `earthaccess.download()` to let the user control display of progress bars. Defaults to true for interactive sessions, otherwise false.
   ([#612](https://github.com/nsidc/earthaccess/issues/612))
+  ([#1065](https://github.com/nsidc/earthaccess/pull/1065))
   ([@Sherwin-14](https://github.com/Sherwin-14))
 - Updated bug and triage label names in bug Issue template.
   ([#998](https://github.com/nsidc/earthaccess/pull/998))
@@ -21,9 +27,15 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 - `GESDISC` should be `GES_DISC` in docstrings.
   ([#1037](https://github.com/nsidc/earthaccess/issues/1037))
   ([@abarciauskas-bgse](https://github.com/abarciauskas-bgse))
+- `open_virtual_mfdataset` now uses `virtualizarr` v2, and `obstore` in place of `fsspec`. Updated Zarr to V3 xref #967.
+  ([#1074](https://github.com/nsidc/earthaccess/issues/1074))
+  ([@owenlittlejohns](https://github.com/owenlittlejohns))
 
 ### Added
 
+- Added `tenacity` to retry downloads up to 3 times with exponential backoff time, replaces #1016
+  ([#481](https://github.com/nsidc/earthaccess/issues/481))
+  ([@betolink](https://github.com/betolink))
 - Add notebook demonstrating workflow with TEMPO Level 3 data as a virtual dataset
   ([#924](https://github.com/nsidc/earthaccess/pull/924))
   ([@danielfromearth](https://github.com/danielfromearth))
@@ -37,12 +49,15 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
   ([@Sherwin-14](https://github.com/Sherwin-14))
   ([@mfisher87](https://github.com/mfisher87))
 - `download` now returns Path consistently.
-  ([#595])(https://github.com/nsidc/earthaccess/issues/595)
+  ([#595])(<https://github.com/nsidc/earthaccess/issues/595>)
   ([@Sherwin-14](https://github.com/Sherwin-14))
 - Users may now authenticate with an existing Earthdata login token with
   environment variable `EARTHDATA_TOKEN`
   ([#484](https://github.com/nsidc/earthaccess/issues/484))
   ([@kgrimes2](https://github.com/kgrimes2))
+- Added top level `status` function to check the statuses of NASA Earthdata services
+  ([#161](https://github.com/nsidc/earthaccess/issues/161))
+  ([@Sherwin-14](https://github.com/Sherwin-14))
 
 ### Removed
 
@@ -56,8 +71,12 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ### Fixed
 
+- Files can be downloaded in the cloud([#1009](https://github.com/nsidc/earthaccess/issues/1009))([betolink](https://github.com/betolink))
 - Corrected Harmony typo in notebooks/Demo.ipynb([#995](https://github.com/nsidc/earthaccess/issues/995))([stelios-c](https://github.com/stelios-c))
 - Resolved an error in virtual dataset tutorial notebook ([#1044](https://github.com/nsidc/earthaccess/issues/1044))([danielfromearth](https://github.com/danielfromearth))
+- Issue when `FileDistributionInformation` did not exist for a collection
+  ([#971](https://github.com/nsidc/earthaccess/pull/971))
+  ([@mike-gangl](https://github.com/mike-gangl/))
 
 ## [v0.14.0] - 2025-02-11
 
@@ -748,9 +767,6 @@ _Conception!_
 - Basic object formatting.
 
 [Unreleased]: https://github.com/nsidc/earthaccess/compare/v0.14.0...HEAD
-[0.14.0]: https://github.com/nsidc/earthaccess/compare/v0.13.0...v0.14.0
-[0.13.0]: https://github.com/nsidc/earthaccess/compare/v0.12.0...v0.13.0
-[0.12.0]: https://github.com/nsidc/earthaccess/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/nsidc/earthaccess/releases/tag/v0.11.0
 [0.10.0]: https://github.com/nsidc/earthaccess/releases/tag/v0.10.0
 [0.9.0]: https://github.com/nsidc/earthaccess/releases/tag/v0.9.0